refactor detector sweep function
